+++ src/actions/hand.js
@@ -1,5 +1,4 @@
 import { moveToArray } from "../utils/help";
-import { shuffleDeck } from "./deck";
 import Action from "../models/action";
 import MenuData from "../models/menu";
 import { Origin } from "../models/enums";
@@ -13,11 +12,6 @@ export const backToTopDeck = (G, ctx, index) => {
 
 export const backToBottomDeck = (G, ctx, index) => {
   moveToArray(G.hand[ctx.playerID], G.deck[ctx.playerID].cards, index, true);
-};
-
-export const backToDeck = (G, ctx, index) => {
-  moveToArray(G.hand[ctx.playerID], G.deck[ctx.playerID].cards, index);
-  shuffleDeck(G, ctx);
 };
 
 export const destroy = (G, ctx, index) => {
@@ -51,7 +45,6 @@ const getHandActionsOnMenu = (event, pos, hand, myID) => {
   let actions = [
     Action("Back to Top of the Deck", backToTopDeck.name, [pos]),
     Action("Back to Bottom of the Deck", backToBottomDeck.name, [pos]),
-    Action("Back to the Deck and Shuffle", backToDeck.name, [pos]),
     Action("SPAWN Face-up", spawnFaceUp.name, [hand, myID, pos]),
     Action("SPAWN Face-down", spawnFaceDown.name, [hand, myID, pos]),
     Action("Destroy", destroy.name, [pos]),